feat: Use normalized url strings for filename.
I download a lot of unicode-happy tracks with insane filenames, and the
"only allow a small subset of special characters" approach fails hard.

E.G: https://soundcloud.com/drwartemal/when-a-fire-starts-to-burn
becomes
"  &   -    tt t .mp3"

This small tweak uses the normalized artist and title from the url
scheme for the filenames instead. The fancy characters are still
preserved in the metadata.